+++ tiny_listener/listener.py
@@ -1,40 +1,56 @@
 import asyncio
 import signal
 from typing import Optional, Dict, Callable, List, Any, Union, Tuple, Type
-from concurrent.futures import CancelledError
 from copy import copy
 
 from .context import Context
 from .routing import Route, _EventHandler, EventHandler, as_handler, Params
 
 
-class NotFound(BaseException):
+class RouteNotFound(BaseException):
+    pass
+
+
+class ContextNotFound(BaseException):
+    pass
+
+
+class DuplicatedCid(BaseException):
     pass
 
 
 class Listener:
-    __default_routes__: List[Route] = []
-    __default_pre_do__: List[EventHandler] = []
-    __default_post_do__: List[EventHandler] = []
-    __default_error_raise__: List[Tuple[EventHandler, Type[BaseException]]] = []
-
     def __init__(self):
         self.loop = asyncio.new_event_loop()
         self.loop.set_exception_handler(self.exception_handler)
         for sig in [signal.SIGINT, signal.SIGTERM]:
             self.loop.add_signal_handler(sig, self.exit)
 
-        self.routes = copy(self.__default_routes__)
-        self._pre_do = copy(self.__default_pre_do__)
-        self._post_do = copy(self.__default_post_do__)
-        self._error_raise = copy(self.__default_error_raise__)
+        self.ctxs: Dict[str, Context] = {}
+        self.routes: List[Route] = []
+        self._pre_do: List[EventHandler] = []
+        self._post_do: List[EventHandler] = []
+        self._error_raise: List[Tuple[EventHandler, Type[BaseException]]] = []
 
-    def new_context(self, cid: Optional[str] = None, **scope: Any) -> Context:
-        return Context(cid, _listener_=self, **scope)
+    def new_ctx(self, cid: str = "__main__", scope: Optional[Dict] = None) -> Context:
+        if self.get_ctx(cid):
+            raise DuplicatedCid(f"Context `{cid}` already exist")
+
+        ctx = Context(listener=self, cid=cid, scope=scope)
+        self.ctxs[cid] = ctx
+        return ctx
+
+    def get_ctx(self, cid: str) -> Optional[Context]:
+        return self.ctxs.get(cid)
+
+
+    # def is_context_exist(self, cid: str) -> bool:
+    #     return self.ctxs.get(cid) is not None
+
 
     @staticmethod
     def exception_handler(loop, context):
-        if isinstance(context.get("exception"), CancelledError):
+        if isinstance(context.get("exception"), asyncio.CancelledError):
             loop.stop()
         else:
             loop.default_exception_handler(context)
@@ -59,12 +75,11 @@ class Listener:
             self,
             name: str,
             cid: Optional[str] = None,
-            timeout: Optional[float] = None,
+            parents_timeout: Optional[float] = None,
             data: Optional[Dict] = None
     ):
-        ctx = self.new_context(cid)
-
         async def _fire():
+            ctx = self.new_ctx(cid)
             params = Params()
             route = None
             for r in self.routes:
@@ -73,10 +88,10 @@ class Listener:
                     route = r
                     params = Params(params)
                     break
-            event = ctx.new_event(name=name, timeout=timeout, route=route, **data or {})
+            event = ctx.new_event(name=name, parents_timeout=parents_timeout, route=route, **data or {})
 
-            try:
-                async with event as exc:
+            async with event as exc:
+                try:
                     if event.route is None:
                         raise NotFound(f"route `{name}` not found")
                     if exc:
@@ -85,10 +100,13 @@ class Listener:
                     res = await route.fn(ctx, event, params)
                     [await fn(ctx, event, params, exc) for fn in self._post_do]
                     return res
-            except BaseException as e:
-                if not self._error_raise:
-                    raise e
-                [await fn(ctx, event, params, e) for fn, exc_cls in self._error_raise if isinstance(e, exc_cls)]
+                except BaseException as e:
+                    if not self._error_raise:
+                        raise e
+                    event.error = e
+                    [await fn(ctx, event, params, e) for fn, exc_cls in self._error_raise if isinstance(e, exc_cls)]
+                    if event.error is not None:
+                        ctx.drop(cid)
 
         return self.loop.create_task(_fire())
 
@@ -110,33 +128,6 @@ class Listener:
         def _decorator(fn: _EventHandler) -> None:
             self.routes.append(Route(path=path, fn=fn, parents=parents, opts=kwargs))
         return _decorator
-
-    @classmethod
-    def default_do(
-            cls,
-            path: str,
-            parents: Union[None, List[str], Callable[[Context], List[str]]] = None,
-            **kwargs: Any
-    ) -> Callable[[_EventHandler], None]:
-        parents = parents or []
-
-        def _decorator(fn: _EventHandler) -> None:
-            cls.__default_routes__.append(Route(path=path, fn=fn, parents=parents, opts=kwargs))
-        return _decorator
-
-    @classmethod
-    def default_pre_do(cls, fn: _EventHandler) -> None:
-        cls.__default_pre_do__.append(as_handler(fn))
-
-    @classmethod
-    def default_post_do(cls, fn: _EventHandler) -> None:
-        cls.__default_post_do__.append(as_handler(fn))
-
-    @classmethod
-    def default_error_raise(cls, exc: Type[BaseException]) -> Callable[[_EventHandler], None]:
-        def f(fn: _EventHandler) -> None:
-            cls.__default_error_raise__.append((as_handler(fn), exc))
-        return f
 
     def __repr__(self) -> str:
         return "{}(listener_count={})".format(self.__class__.__name__, len(self.routes))